remove events and rename to law_policy
