+++ app/data_migrations/populate_geography.py
@@ -4,7 +4,7 @@ from typing import cast
 from slugify import slugify
 from sqlalchemy.orm import Session
 
-from app.db.models.lawpolicy import Geography
+from app.db.models.law_policy import Geography
 
 from .utils import has_rows, load_tree
