+++ app/db/crud/document.py
@@ -59,7 +59,7 @@ from app.db.models.deprecated import (
     Source,
 )
 from app.db.models.deprecated import DocumentType
-from app.db.models.lawpolicy import Geography
+from app.db.models.law_policy import Geography
 
 _LOGGER = logging.getLogger(__file__)
