+++ app/db/models/__init__.py
@@ -1,5 +1,5 @@
 from app.db.models import deprecated
 from app.db.models import document
 from app.db.models import app
-from app.db.models import lawpolicy
+from app.db.models import law_policy
 from app.db.session import Base
