+++ app/db/models/law_policy/slug.py
