+++ tests/core/db/test_crud.py
@@ -19,7 +19,7 @@ from app.db.models.deprecated import (
 from app.api.api_v1.schemas.document import DocumentCreateRequest
 from app.db.crud.document import create_document, UnknownMetadataError
 from app.db.models.deprecated import DocumentType
-from app.db.models.lawpolicy import Geography
+from app.db.models.law_policy import Geography
 
 
 def test_create_documents(client, superuser_token_headers, test_db):
