+++ tests/routes/test_data_fixtures.py
@@ -2,7 +2,7 @@ from dateutil.parser import parse
 import pytest
 from app.db.models.deprecated import DocumentType
 from app.db.models.deprecated import Category, Document, Source, Event
-from app.db.models.lawpolicy import Geography
+from app.db.models.law_policy import Geography
 from tests.utils import json_serialize
 
 template_doc = {
