+++ tests/routes/test_documents.py
@@ -19,7 +19,7 @@ from app.db.crud.document import (
     get_document_detail,
     get_postfix_map,
 )
-from app.db.models.lawpolicy import Geography
+from app.db.models.law_policy import Geography
 from app.db.models.deprecated import DocumentType
